+++ lib/safire/client_config_builder.rb
@@ -50,6 +50,15 @@ module Safire
       self
     end
 
+    # Sets the leaf-first X.509 certificate chain used for UDAP signing.
+    #
+    # @param certificate_chain [Array<String, OpenSSL::X509::Certificate>, nil]
+    # @return [self]
+    def certificate_chain(certificate_chain)
+      @config[:certificate_chain] = certificate_chain
+      self
+    end
+
     def kid(kid)
       @config[:kid] = kid
       self
@@ -62,6 +71,15 @@ module Safire
 
     def jwks_uri(jwks_uri)
       @config[:jwks_uri] = jwks_uri
+      self
+    end
+
+    # Allows HTTP loopback URIs for local development.
+    #
+    # @param enabled [Boolean] set to +true+ only when a local server cannot terminate TLS
+    # @return [self]
+    def allow_insecure_localhost(enabled: true)
+      @config[:allow_insecure_localhost] = enabled
       self
     end